requirements.txt
flask
hydra-core
jupyter
librosa
matplotlib
numpy
optuna
pytest
python-levenshtein
scipy
scikit-learn
sox
torch
torchaudio
torchelastic
tqdm
wget
waitress